#163426515 Deploy base app to heroku (#7)
chore: deploy base app to heroku
