+++ .eslintrc.json
@@ -14,8 +14,8 @@
     "no-param-reassign": 0,
     "comma-dangle": 0,
     "curly": ["error", "multi-line"],
-    "import/no-unresolved": [2, { commonjs: true }],
-    "no-shadow": ["error", { allow: ["req", "res", "err"] }],
+    "import/no-unresolved": [2, { "commonjs": true }],
+    "no-shadow": ["error", { "allow": ["req", "res", "err"] }],
     "valid-jsdoc": [
       "error",
       {
@@ -36,4 +36,4 @@
       }
     ]
   }
-};
+}
