+++ .gitignore
@@ -5,7 +5,8 @@ logs
 .env 
 .nyc_output
 npm-debug.log*
-
+build
+package-lock.json
 # Runtime data
 pids
 *.pid
