+++ .travis.yml
@@ -1,11 +1,15 @@
 language: node_js
 node_js:
-  - "node"
+  - "stable"
+env:
+  - global
 services:
   - postgresql
+install:
+  - npm install
 before_script:
   - psql -c 'create database travis_ci_test;' -U postgres
 script:
-  - npm start
+  - npm test
 after_success:
   - npm run coverage
