+++ README.md
@@ -409,3 +409,8 @@ No additional parameters required
 ### Get Tags
 
 `GET /api/tags`
+
+
+<hr>
+
+You can access the app here https://indiana-ah-staging.herokuapp.com/
