+++ package.json
@@ -1,50 +1,60 @@
 {
-  "name": "express-authorshaven",
-  "version": "1.0.0",
-  "description": "A Social platform for the creative at heart",
-  "main": "index.js",
-  "scripts": {
-    "test": "cross-env NODE_ENV=test nyc mocha server/test/**/*.js --require @babel/register --timeout 6000 --exit",
-    "start": "nodemon --exec babel-node index.js",
-    "coverage": "nyc report --reporter=text-lcov | coveralls && nyc report --reporter=lcov"
-  },
-  "author": "Andela Simulations Programme",
-  "license": "MIT",
-  "dependencies": {
-    "@babel/register": "^7.0.0",
-    "body-parser": "^1.18.3",
-    "chai": "^4.2.0",
-    "cors": "^2.8.4",
-    "cross-env": "^5.2.0",
-    "dotenv": "^6.0.0",
-    "mocha": "^5.2.0",
-    "nyc": "^13.1.0",
-    "ejs": "^2.6.1",
-    "errorhandler": "^1.5.0",
-    "express": "^4.16.3",
-    "express-jwt": "^5.3.1",
-    "express-session": "^1.15.6",
-    "jsonwebtoken": "^8.3.0",
-    "method-override": "^2.3.10",
-    "methods": "^1.1.2",
-    "morgan": "^1.9.1",
-    "passport": "^0.4.0",
-    "passport-local": "^1.0.0",
-    "pg": "^7.8.0",
-    "pg-hstore": "^2.3.2",
-    "request": "^2.87.0"
-  },
-  "devDependencies": {
-    "eslint": "^5.12.1",
-    "eslint-config-airbnb-base": "^13.1.0",
-    "eslint-plugin-import": "^2.15.0",
-    "sequelize": "^4.42.0",
-    "slug": "^0.9.1",
-    "underscore": "^1.9.1",
-    "@babel/cli": "^7.2.3",
-    "@babel/core": "^7.2.2",
-    "@babel/node": "^7.2.2",
-    "@babel/preset-env": "^7.3.1",
-    "nodemon": "^1.18.9"
-  }
+	"name": "express-authorshaven",
+	"version": "1.0.0",
+	"description": "A Social platform for the creative at heart",
+	"main": "index.js",
+	"scripts": {
+		"test": "cross-env NODE_ENV=test nyc mocha server/test/**/*.js --require @babel/register --timeout 6000 --exit",
+		"clean": "rm -rf build",
+		"build": "npm run clean && mkdir build && cd build && mkdir server && babel ./server  -s -d server",
+		"start": "npm run build && node --require @babel/register ./server/index.js",
+		"dev": "nodemon --exec babel-node server/index.js",
+		"coverage": "nyc report --reporter=text-lcov | coveralls && nyc report --reporter=lcov"
+	},
+	"engines": {
+		"node": "10.x"
+	},
+	"author": "Andela Simulations Programme",
+	"license": "MIT",
+	"dependencies": {
+		"@babel/cli": "^7.2.3",
+		"@babel/core": "^7.2.2",
+		"@babel/preset-env": "^7.3.1",
+		"@babel/register": "^7.0.0",
+		"babel-core": "^7.0.0-bridge.0",
+		"babel-jest": "^24.0.0",
+		"body-parser": "^1.18.3",
+		"chai": "^4.2.0",
+		"cors": "^2.8.4",
+		"cross-env": "^5.2.0",
+		"dotenv": "^6.0.0",
+		"ejs": "^2.6.1",
+		"errorhandler": "^1.5.0",
+		"express": "^4.16.3",
+		"express-jwt": "^5.3.1",
+		"express-session": "^1.15.6",
+		"jsonwebtoken": "^8.3.0",
+		"method-override": "^2.3.10",
+		"methods": "^1.1.2",
+		"mocha": "^5.2.0",
+		"morgan": "^1.9.1",
+		"nyc": "^13.1.0",
+		"passport": "^0.4.0",
+		"passport-local": "^1.0.0",
+		"pg": "^7.8.0",
+		"pg-hstore": "^2.3.2",
+		"regenerator-runtime": "^0.13.1",
+		"request": "^2.87.0",
+		"slug": "^0.9.3",
+		"underscore": "^1.9.1"
+	},
+	"devDependencies": {
+		"eslint": "^5.12.1",
+		"eslint-config-airbnb-base": "^13.1.0",
+		"eslint-plugin-import": "^2.15.0",
+		"nodemon": "^1.18.9",
+		"sequelize": "^4.42.0",
+		"slug": "^0.9.1",
+		"underscore": "^1.9.1"
+	}
 }
